requirements.txt
kafka-python==2.0.2
scikit-learn==1.0.2
joblib==1.1.0
numpy==1.22.3
flask==2.1.1
loguru==0.5.3
python-dotenv==0.19.2
